Pass CI

+++ src/strategies/prioritise_empty.rs
@@ -42,7 +42,7 @@ impl Strategy for PrioritiseEmptyStrategy {
 
             let new_pos = pos + dir;
 
-            match grid[usize::from(new_pos.y) * usize::from(width) + usize::from(new_pos.x)] {
+            match grid[new_pos.y * width + new_pos.x] {
                 Cell::Empty => priority_dirs.push(dir),
                 Cell::Player(player_id) | Cell::PlayerClaimed(player_id) => {
                     if player_id == id {
